Move file-reading logic out of tasks, put in shared util
